% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biobroom.R
\docType{package}
\name{biobroom}
\alias{biobroom}
\alias{biobroom-package}
\title{Convert Bioconductor Object into Tidy Data Frames}
\description{
This package contains methods for converting standard objects
constructed by bioinformatics packages, especially those in Bioconductor, and
converting them to tidy data. It thus serves as a complement to the broom
package, and follows the same the tidy, augment, glance division of tidying
methods. Tidying data makes it easy to recombine, reshape and visualize
bioinformatics analyses.
}

