% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ExpressionSet_tidiers}
\alias{ExpressionSet_tidiers}
\title{Tidying methods for Biobase's ExpressionSet objects}
\usage{
\S4method{tidy}{ExpressionSet}(x, addPheno = FALSE, ...)
}
\arguments{
\item{x}{ExpressionSet object}

\item{addPheno}{whether columns should be included in the tidied output
for those in the ExpressionSet's phenoData}

\item{...}{extra arguments (not used)}
}
\description{
Tidying methods for Biobase's ExpressionSet objects
}
\details{
\code{addPheno=TRUE} adds columns that are redundant (since they
add per-sample information to a per-sample-per-gene data frame), but that
are useful for some kinds of graphs and analyses.
}
\examples{
data(hammer)
head(tidy(hammer))
head(tidy(hammer, addPheno=TRUE))
}

