% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{finish}
\alias{finish}
\title{Finish a tidied dataset and format for returning}
\usage{
finish(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a data-frame-like object, such as a tbl_df,
tbl_dt, or data.table, whose class is determined by the
biobroom.return option
}
\description{
Each tidy and augment function calls this function before
returning. It ensures the object has no rownames, then turns the
return object into a tbl_df, tbl_dt, or whatever form is requested
using the global biobroom.return option (default tbl_df)
}

