% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GRanges_tidiers.R
\name{GRanges_tidiers}
\alias{GRanges_tidiers}
\alias{tidy.GRanges}
\title{Tidying methods for GRanges and GRangesList objects.}
\usage{
\method{tidy}{GRanges}(x)
}
\arguments{
\item{x}{GRanges or GRangesList object}
}
\value{
All tidying methods return a \code{data.frame} without rownames.

\code{tidy} returns one row for each range, which contains
\item{start} of the range
\item{end} of the range
\item{width} (or length) of the range
\item{names} of the range
\item{strand}
\item{seqname} Name of the sequence from which the range comes (usually the chromosome)
\item{metadata} Any included metadata, (ie, score, GC content)

For \code{GRangesList}, there will also be a column representing which group the ranges comes from.

\code{glance} returns a \code{data.frame} with the number of ranges, the number of sequences, and the number of groups (if applicable).
}
\description{
Tidying methods for GRanges and GRangesList objects.
}
\examples{
source("http://bioconductor.org/biocLite.R")
biocLite("GenomicRanges")
library(GenomicRanges)
#toy data
gr <-
 GRanges(seqnames =
           Rle(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
         ranges =
           IRanges(1:10, end = 7:16, names = head(letters, 10)),
         strand =
           Rle(strand(c("-", "+", "*", "+", "-")),
               c(1, 2, 2, 3, 2)),
         score = 1:10,
         GC = seq(1, 0, length=10))

#make a GRangesList from the toy GRanges object
sp <- split(gr, rep(1:2, each=5))


tidy(gr)
tidy(sp)
glance(gr)
glance(sp)
}

