% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_helpers.R
\name{makeExampleDataSet}
\alias{makeExampleDataSet}
\title{Makes example datasets for different objects we tidy.}
\usage{
makeExampleDataSet(type = "DESeq2", n = 1000, m = 12, betaSD = 0,
  interceptMean = 4, interceptSD = 2, dispMeanRel = function(x) 4/x + 0.1,
  sizeFactors = rep(1, m))
}
\arguments{
\item{type}{The type of object to output. So far we can build objects for
\itemize{
  \item{\code{"DESeq2"}}{
      A \code{\link[DESeq2]{DESeqDataSet}}
  }
  \item{\code{"edgeR"}}{
      A \code{\link[edgeR]{DGEList}}
  }
  \item{\code{"limma"}}{
      An \code{\link[limma]{EList}}
  }
  \item{\code{"voom"}}{
      An \code{\link[limma]{EList}} which results from passing a
      \code{\link[edgeR]{DGEList}} through \code{\link[limma]{voom}}
  }
  \item{\code{"voomWithQualityWeights"}}{
      Same as voom, but run through
      \code{\link[limma]{voomWithQualityWeights}}
  }
  \item{\code{"ExpressionSet"}}{
    An \code{\link[Biobase]{ExpressionSet}}
  }
}}
}
\value{
The desired output object
}
\description{
This is simply a wrapper to the \code{\link[DESeq2]{makeExampleDESeqDataSet}}
function, which then transforms the output to the appropriate type.
}

