% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{DESeqDataSet_tidiers}
\alias{DESeqDataSet_tidiers}
\title{Tidying methods for DESeq2 DESeqDataSet objects}
\usage{
\S4method{tidy}{DESeqDataSet}(x, colData = FALSE, ...)
}
\arguments{
\item{x}{DESeqDataSet object}

\item{colData}{whether colData should be included in the tidied output
for those in the DESeqDataSet object}

\item{...}{extra arguments (not used)}
}
\description{
Tidying methods for DESeq2 DESeqDataSet objects
}
\details{
\code{colDat=TRUE} adds covariates from colData to the data frame.
}
\examples{
# From DESeq2 documentation

library("DESeq2")
dds <- makeExampleDESeqDataSet()

xx <- tidy(dds)

# With covariates
xx <- tidy(dds, colData=TRUE)

# Filter data for counts >20
xx \%>\% filter(counts > 20)
}

