% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{DESeqDataSet_tidiers}
\alias{DESeqDataSet_tidiers}
\title{Tidying methods for Biobase's DESeqDataSet objects}
\usage{
\S4method{tidy}{DESeqDataSet}(x, colData = FALSE, ...)
}
\arguments{
\item{x}{ExpressionSet object}

\item{colData}{whether colData should be included in the tidied output
for those in the DESeqDataSet object}

\item{...}{extra arguments (not used)}
}
\description{
Tidying methods for Biobase's DESeqDataSet objects
}
\details{
\code{colDat=TRUE} adds columns that are redundant (since they
add per-sample information to a per-sample-per-gene data frame), but that
are useful for some kinds of graphs and analyses.
}
\examples{
# Nothing yet
}

