% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeR_tidiers.R
\name{edgeR_tidiers}
\alias{edgeR_tidiers}
\alias{glance.DGEExact}
\alias{tidy.DGEExact}
\title{Tidiers for edgeR's differential expression objects}
\usage{
\method{tidy}{DGEExact}(x, ...)

\method{glance}{DGEExact}(x, alpha = 0.05, p.adjust.method = "fdr", ...)
}
\arguments{
\item{x}{DGEExact object}

\item{...}{extra arguments (not used)}

\item{alpha}{Confidence level to test for significance}

\item{p.adjust.method}{Method for adjusting p-values to determine
significance; can be any in p.adjust.methods}
}
\value{
\code{glance} returns one row with the columns
  \item{significant}{number of significant genes using desired adjustment
  method and confidence level}
  \item{comparison}{The pair of groups compared by edgeR, delimited by /}
}
\description{
Tidy, augment and glance methods for turning edgeR objects into tidy data
frames, where each row represents one observation and each column represents
one column.
}
\examples{
if (require("edgeR")) {
    library(Biobase)
    data(hammer)
    hammer.counts <- exprs(hammer)[, 1:4]
    hammer.treatment <- phenoData(hammer)$protocol[1:4]

    y <- DGEList(counts=hammer.counts,group=hammer.treatment)
    y <- calcNormFactors(y)
    y <- estimateCommonDisp(y)
    y <- estimateTagwiseDisp(y)
    et <- exactTest(y)

    head(tidy(et))
    head(glance(et))
}
}

