% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExpressionSet_tidiers.R
\docType{methods}
\name{ExpressionSet_tidiers}
\alias{ExpressionSet_tidiers}
\title{Tidying methods for Biobase's ExpressionSet objects}
\usage{
\S4method{tidy}{ExpressionSet}(x, addPheno = FALSE, ...)
}
\arguments{
\item{x}{ExpressionSet object}

\item{addPheno}{whether columns should be included in the tidied output
for those in the ExpressionSet's phenoData}

\item{...}{extra arguments (not used)}
}
\description{
Tidying methods for Biobase's ExpressionSet objects
}
\details{
\code{addPheno=TRUE} adds columns that are redundant (since they
add per-sample information to a per-sample-per-gene data frame), but that
are useful for some kinds of graphs and analyses.
}
\examples{
# import ExpressionSet object
data(hammer)

# Use tidy to extract genes, sample ids and measured value
tidy(hammer)
# add phenoType data
tidy(hammer, addPheno=TRUE)

# To look at the data with counts > 30
library("dplyr")

tt \%>\% filter(value > 30)
# to further filter for sample ids
tt \%>\% filter(value > 30 & sample.id =="SRX020102")
}

