% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hammer-data.R
\docType{data}
\name{hammer}
\alias{hammer}
\title{ExpressionSet results from Hammer et al 2010}
\format{\preformatted{Formal class 'ExpressionSet' [package "Biobase"] with 7 slots
  ..@ experimentData   :Formal class 'MIAME' [package "Biobase"] with 13 slots
  .. .. ..@ name             : chr ""
  .. .. ..@ lab              : chr ""
  .. .. ..@ contact          : chr ""
  .. .. ..@ title            : chr ""
  .. .. ..@ abstract         : chr ""
  .. .. ..@ url              : chr ""
  .. .. ..@ pubMedIds        : chr ""
  .. .. ..@ samples          : list()
  .. .. ..@ hybridizations   : list()
  .. .. ..@ normControls     : list()
  .. .. ..@ preprocessing    : list()
  .. .. ..@ other            : list()
  .. .. ..@ .__classVersion__:Formal class 'Versions' [package "Biobase"] with 1 slot
  .. .. .. .. ..@ .Data:List of 2
  .. .. .. .. .. ..$ : int [1:3] 1 0 0
  .. .. .. .. .. ..$ : int [1:3] 1 1 0
  ..@ assayData        :<environment: 0x389fd18> 
  ..@ phenoData        :Formal class 'AnnotatedDataFrame' [package "Biobase"] with 4 slots
  .. .. ..@ varMetadata      :'data.frame':	5 obs. of  1 variable:
  .. .. .. ..$ labelDescription: chr [1:5] NA NA NA NA ...
  .. .. ..@ data             :'data.frame':	8 obs. of  5 variables:
  .. .. .. ..$ sample.id    : Factor w/ 8 levels "SRX020088-90",..: 5 6 7 8 2 1 3 4
  .. .. .. ..$ num.tech.reps: num [1:8] 1 2 1 2 1 2 1 2
  .. .. .. ..$ protocol     : Factor w/ 2 levels "control","L5 SNL": 1 1 2 2 1 1 2 2
  .. .. .. ..$ strain       : Factor w/ 1 level "Sprague Dawley": 1 1 1 1 1 1 1 1
  .. .. .. ..$ Time         : Factor w/ 3 levels "2months","2 months",..: 2 2 2 1 3 3 3 3
  .. .. ..@ dimLabels        : chr [1:2] "sampleNames" "sampleColumns"
  .. .. ..@ .__classVersion__:Formal class 'Versions' [package "Biobase"] with 1 slot
  .. .. .. .. ..@ .Data:List of 1
  .. .. .. .. .. ..$ : int [1:3] 1 1 0
  ..@ featureData      :Formal class 'AnnotatedDataFrame' [package "Biobase"] with 4 slots
  .. .. ..@ varMetadata      :'data.frame':	1 obs. of  1 variable:
  .. .. .. ..$ labelDescription: chr NA
  .. .. ..@ data             :'data.frame':	29516 obs. of  1 variable:
  .. .. .. ..$ gene: Factor w/ 29516 levels "ENSRNOG00000000001",..: 1 2 3 4 5 6 7 8 9 10 ...
  .. .. ..@ dimLabels        : chr [1:2] "featureNames" "featureColumns"
  .. .. ..@ .__classVersion__:Formal class 'Versions' [package "Biobase"] with 1 slot
  .. .. .. .. ..@ .Data:List of 1
  .. .. .. .. .. ..$ : int [1:3] 1 1 0
  ..@ annotation       : chr(0) 
  ..@ protocolData     :Formal class 'AnnotatedDataFrame' [package "Biobase"] with 4 slots
  .. .. ..@ varMetadata      :'data.frame':	0 obs. of  1 variable:
  .. .. .. ..$ labelDescription: chr(0) 
  .. .. ..@ data             :'data.frame':	8 obs. of  0 variables
  .. .. ..@ dimLabels        : chr [1:2] "sampleNames" "sampleColumns"
  .. .. ..@ .__classVersion__:Formal class 'Versions' [package "Biobase"] with 1 slot
  .. .. .. .. ..@ .Data:List of 1
  .. .. .. .. .. ..$ : int [1:3] 1 1 0
  ..@ .__classVersion__:Formal class 'Versions' [package "Biobase"] with 1 slot
  .. .. ..@ .Data:List of 4
  .. .. .. ..$ : int [1:3] 2 13 0
  .. .. .. ..$ : int [1:3] 2 12 1
  .. .. .. ..$ : int [1:3] 1 3 0
  .. .. .. ..$ : int [1:3] 1 0 0
}}
\usage{
hammer
}
\value{
ExpressionSet
}
\description{
An ExpressionSet containing the results of the Hammer et al
2010 RNA-Seq study on the nervous system of rats (Hammer et al 2010).

This was downloaded from the ReCount database of analysis-ready RNA-Seq
datasets (Frazee et al 2011).

Hammer, P., Banck, M. S., Amberg, R., Wang, C., Petznick, G., Luo, S.,
Khrebtukova, I., Schroth, G. P., Beyerlein, P., and Beutler, A. S. (2010).
mRNA-seq with agnostic splice site discovery for nervous system
transcriptomics tested in chronic pain. Genome research, 20(6), 847-860.
http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2877581/

Frazee, A. C., Langmead, B., and Leek, J. T. (2011). ReCount: a
multi-experiment resource of analysis-ready RNA-seq gene count datasets.
BMC Bioinformatics, 12, 449. http://bowtie-bio.sourceforge.net/recount/
}
\keyword{datasets}

