% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/limma_tidiers.R
\name{tidy.MAList}
\alias{tidy.MAList}
\title{Tidying method for an MA list}
\usage{
\method{tidy}{MAList}(x, ...)
}
\arguments{
\item{x}{\code{MAList} class from the limma package.}

\item{...}{Currently no additional arguments are necessary.}
}
\description{
Tidying method for an MA list
}

